Refactor formatter
