+++ lib/compare_linker/formatter/text.rb
@@ -2,7 +2,7 @@ require "ostruct"
 
 class CompareLinker
   class Formatter
-    class Text
+    class Text < Base
       def format(gem_info)
         g = OpenStruct.new(gem_info)
 
@@ -24,17 +24,6 @@ class CompareLinker
         end
 
         text
-      end
-
-      private
-
-      def downgrade?(old_ver, new_ver, old_tag, new_tag)
-        Gem::Version.new(new_ver) < Gem::Version.new(old_ver) ||
-          (old_tag && new_tag && Gem::Version.new(to_ver(new_tag)) < Gem::Version.new(to_ver(old_tag)))
-      end
-
-      def to_ver(tag)
-        tag.sub(/\Av/, '')
       end
     end
   end
